fix: fix uni formula collab
